Dataset preparation

+++ scripts/prepare-data.sh
@@ -4,5 +4,9 @@ cd "$(dirname "$0")/.."
 if [ ! -d "data" ]; then
     mkdir -p data
 fi
-curl -L -o data/MDC_dataset.tar.gz https://github.com/pappd90/mdc_dataset/raw/main/data/MDC_dataset.tar.gz
+if [ ! -f "data/MDC_dataset.tar.gz" ]; then
+    curl -L -o data/MDC_dataset.tar.gz https://github.com/pappd90/mdc_dataset/raw/main/data/MDC_dataset.tar.gz
+fi
 tar -xzf data/MDC_dataset.tar.gz -C data
+
+python scripts/prepare-dataset.py
